refactor reusable components
